# Conflicts:
#	build/web/sidebar.jsp
#	nbproject/private/private.xml
#	web/sidebar.jsp

+++ src/java/com/model/pdb_owner/PDBVersionDB.java
@@ -660,14 +660,16 @@ public class PDBVersionDB {
         }
         return row;
     }
-    public static void deletePDBVersion_Group(int pdbversion_id) throws SQLException{
+    public static void deletePDBVersion_Group(int pdbversion_id, String action_type) throws SQLException{
         Connection connection = null;
         PreparedStatement preparedStatement = null;
-        
-        connection = ConnectionConfiguration.getConnection();
-        preparedStatement = connection.prepareStatement("delete from pdbversion_group where pdbversion_id="+pdbversion_id+" AND id NOT IN ("+StringUtils.join(GlobalDataStore.globalData, ',')+")");
-        preparedStatement.executeUpdate();
-
+        System.out.println("deletepdbversiongroup"+GlobalDataStore.globalData);
+        System.out.println("action_type"+action_type);
+        if(action_type.equals("update")){
+            connection = ConnectionConfiguration.getConnection();
+            preparedStatement = connection.prepareStatement("delete from pdbversion_group where pdbversion_id="+pdbversion_id+" AND id NOT IN ("+StringUtils.join(GlobalDataStore.globalData, ',')+")");
+            preparedStatement.executeUpdate();
+        }
         GlobalDataStore.globalData.clear();
     }
 }
